Modularize summary generation and fix authentication issues
- Create lib/summary-generator.ts with reusable summary generation logic
- Refactor app/api/session/[id]/summary/route.ts to use modular approach
- Update app/api/session/[id]/end/route.ts to use the same module
- Fix React import in AuthProvider to resolve useState hook errors
- Ensure consistent summary generation across both endpoints
- All summary generation now uses the same proven logic and error handling
